Move github-update helper out of tap root
Homebrew treats top-level Ruby files in this tap as formula files.
Keeping github-update.rb at the repository root causes brew update-report
to try instantiating GithubUpdate as a Formula and crash with an arity error.

Move the helper into lib/ and update formula/test references.
